build-ai-agent/requirements.txt
langchain==0.1.0
openai==1.10.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain-community==0.0.13
langchain-openai==0.0.5
